+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Redis based implementations of this for the types found in Hikari's gateway cache interface can be found
   in [hikari-sake](https://github.com/FasterSpeeding/Sake) \>=v1.0.1a1 (exposed by
   `RedisResource.add_to_tanjun`).
+- Scheduled callback interface and interval implementation to the standard component implementation.
 
 ### Removed
 - `BaseConverter.convert` in-favour of having each standard converter directly implement `__call__`.